Dynamic configuration: Refactor
