+++ src/LuigisBoxBundle/DTO/ConfigDTO.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace Answear\LuigisBoxBundle\DTO;
 
+use Answear\LuigisBoxBundle\DependencyInjection\Configuration;
+
 class ConfigDTO
 {
     private const MAX_SEARCH_CACHE_TTL = 300;
@@ -46,11 +48,11 @@ class ConfigDTO
     public function __construct(
         string $publicKey,
         string $privateKey,
-        string $host = 'https://live.luigisbox.com',
-        float $connectionTimeout = 4.0,
-        float $requestTimeout = 10.0,
-        float $searchTimeout = 6.0,
-        int $searchCacheTtl = 0
+        string $host = Configuration::HOST,
+        float $connectionTimeout = Configuration::CONNECTION_TIMEOUT,
+        float $requestTimeout = Configuration::REQUEST_TIMEOUT,
+        float $searchTimeout = Configuration::SEARCH_TIMEOUT,
+        int $searchCacheTtl = Configuration::SEARCH_CACHE_TIMEOUT
     ) {
         if ($searchCacheTtl < 0) {
             throw new \InvalidArgumentException('searchCacheTtl cannot be negative.');
